// src/api/forge/SettingSetRequestModel.ts
/**
 * Edit Setting request model
 * @export
 * @interface SettingSetRequestModel
 */
export interface SettingSetRequestModel {
  /**
   * setting key
   * @type {string}
   * @memberof SettingSetRequestModel
   */
  key: string;
  /**
   * new value
   * @type {string}
   * @memberof SettingSetRequestModel
   */
  value: string;
}